make SphericalGeometry a module
